feature: search email and password
